CRCBS

CRCBS.AStar

CRCBS.AbstractCostModel

CRCBS.AbstractLowLevelEnv

CRCBS.CBSSolver

CRCBS.CompositeCostModel

CRCBS.ConflictTable

CRCBS.ConstraintTreeNode

CRCBS.DeadlineCost

CRCBS.DiscreteConstraintTable

CRCBS.EnvDistanceHeuristic

CRCBS.FeatureExtractor

CRCBS.FullCostModel

CRCBS.HardConflictTable

CRCBS.LowLevelCostModel

CRCBS.LowLevelSolution

CRCBS.MAPF

CRCBS.MetaAgentCBS_Solver

CRCBS.MetaCost

CRCBS.MetaCostModel

CRCBS.MetaSolution

CRCBS.MultiDeadlineCost

CRCBS.MultiStagePerfectHeuristic

CRCBS.NullHeuristic

CRCBS.PIBTCache

CRCBS.PIBTPlanner

CRCBS.PIBTReservationTable

CRCBS.Path

CRCBS.PathNode

CRCBS.PerfectHeuristic

CRCBS.ReservationTable

CRCBS.ResourceReservation

CRCBS.SoftConflictTable

CRCBS.SoftConflictTable

CRCBS.SolverException

CRCBS.SolverLogger

CRCBS.SolverWrapper

CRCBS.a_star!

CRCBS.a_star_impl!

CRCBS.accumulate_cost

CRCBS.action_space_trait

CRCBS.add_conflict!

CRCBS.add_constraint!

CRCBS.add_constraint!

CRCBS.add_fat_path_to_table!

CRCBS.add_heuristic_cost

CRCBS.add_to_path!

CRCBS.aggregate_costs

CRCBS.aggregate_costs_meta

CRCBS.build_env

CRCBS.cbs!

CRCBS.cbs_branch!

CRCBS.cbs_bypass!

CRCBS.cbs_update_conflict_table!

CRCBS.check_termination_criteria

CRCBS.combine_agents!

CRCBS.compute_path_cost

CRCBS.construct_config_dataframe

CRCBS.construct_empty_lookup_table

CRCBS.construct_results_dataframe

CRCBS.convert_to_vertex_lists

CRCBS.count_conflicts

CRCBS.create_reservations

CRCBS.create_reservations

CRCBS.default_solution

CRCBS.deserialize

CRCBS.detect_action_conflict

CRCBS.detect_action_conflict

CRCBS.detect_conflicts!

CRCBS.detect_conflicts!

CRCBS.detect_conflicts!

CRCBS.detect_conflicts!

CRCBS.detect_state_conflict

CRCBS.detect_state_conflict

CRCBS.extend_path

CRCBS.extend_path!

CRCBS.extract_feature

CRCBS.generate_constraints_from_conflict

CRCBS.get_agent_idxs

CRCBS.get_conflict_index

CRCBS.get_conflicting_paths

CRCBS.get_conflicts

CRCBS.get_constraints

CRCBS.get_cost_model

CRCBS.get_fat_path

CRCBS.get_fat_path_cost

CRCBS.get_heuristic_cost

CRCBS.get_infeasible_cost

CRCBS.get_infeasible_solution

CRCBS.get_initial_cost

CRCBS.get_initial_solution

CRCBS.get_level_set_nodes

CRCBS.get_next_conflict

CRCBS.get_next_state

CRCBS.get_path_cost

CRCBS.get_path_node

CRCBS.get_possible_actions

CRCBS.get_transition_cost

CRCBS.handle_solver_exception

CRCBS.hard_reset_solver!

CRCBS.has_constraint

CRCBS.init_dataframe

CRCBS.init_mapf_2

CRCBS.init_mapf_3

CRCBS.init_mapf_4

CRCBS.init_mapf_5

CRCBS.initialize_child_search_node

CRCBS.initialize_root_node

CRCBS.is_available

CRCBS.is_consistent

CRCBS.is_goal

CRCBS.is_valid

CRCBS.load_problem

CRCBS.low_level_search!

CRCBS.num_actions

CRCBS.num_states

CRCBS.partially_set_path!

CRCBS.pibt_priority_law

CRCBS.pibt_step!

CRCBS.populate_soft_lookup_table!

CRCBS.recompute_cost

CRCBS.reconstruct_path!

CRCBS.remove_constraint!

CRCBS.reserve!

CRCBS.reserved_by

CRCBS.reset_solver!

CRCBS.reset_undecided!

CRCBS.run_profiling

CRCBS.search_constraints

CRCBS.serialize

CRCBS.set_constraint!

CRCBS.solve!

CRCBS.sorted_actions

CRCBS.state_space_trait

CRCBS.states_match

CRCBS.time_to_deadline

CRCBS.trim_path!

CRCBS.trim_solution!

CRCBS.update_fat_path_conflict_table!

CRCBS.violates_constraints

CRCBS.wait

CRCBS.write_results

GraphUtils.get_graph

CRCBS.AStar — Type

AStar

A* Path Planner. Fields:

logger

replan : if true, planner will replan with an empty conflict table following timeout.

CRCBS.AbstractCostModel — Type

`AbstractCostModel{T}`

CRCBS.AbstractLowLevelEnv — Type

`AbstractLowLevelEnv{S,A,C}`

Defines a prototype environment for low level search (searching for a path
for a single agent).

`S` is the State type, `A` is the action type, and `C` is the cost type. All
three must be default constructible (i.e. you can call `S()`, `A()` and `C()`
without throwing errors)

In general, a concrete subtype of `AbstractLowLevelEnv` may include a graph
whose edges are traversed by agents.

CRCBS.CBSSolver — Type

CBSSolver

Path planner that employs Conflict-Based Search.

CRCBS.CompositeCostModel — Type

`CompositeCost{T}`

CRCBS.ConflictTable — Type

A lookup table to store all conflicts that have been detected

CRCBS.ConstraintTreeNode — Type

A node of a constraint tree. Each node has a set of constraints, a candidate
solution (set of robot paths), and a cost

CRCBS.DeadlineCost — Type

`DeadlineCost`

`DeadlineCost` is identical to `TravelTime`, except for the behavior of
`add_heuristic_cost`.

add_heuristic_cost: `c = max(0.0, t + Δt - deadline)`

CRCBS.DiscreteConstraintTable — Type

DiscreteStateTable

Stores constraints for a discrete state space

CRCBS.EnvDistanceHeuristic — Type

EnvDistanceHeuristic

A convenience struct that allows for the distance matrix to be stored in env instead of the heuristic struct.

CRCBS.FeatureExtractor — Type

FeatureExtractor

Abstract type for features that may be reported about the solution to a PC-TAPF or sequential task assignment problem

CRCBS.FullCostModel — Type

`FullCostModel{F,T,M<:AbstractCostModel{T}} <: AbstractCostModel{T}`

The `FullCostModel` defines all the behavior required for running CBS-based
algorithms.

Elements:
- f::F must be callable, and defines how a vector of path costs (e.g., the
    paths of a solution) should be combined into a single cost that reflects
    the cost of the entire path group together
- model::M<:AbstractCostModel{T} defines how the cost is computed during low
level search (when individual paths are being compared against each other).

CRCBS.HardConflictTable — Type

`HardConflictTable`

Stores a lookup table of planned paths for all agents.
When agent `i` queries the table, `table.paths[i]` (the existing path plan
for agent `i`) must be subtracted so that the agent does not try to avoid
conflicts with itself.

CRCBS.LowLevelCostModel — Type

`LowLevelCostModel{C}`

The low level cost model defines the objective to be optimized by the
solver at the low level. An optimal low level solver will return a path if a
feasible path exists) of minimal cost under the objective specified by the
associated LowLevelCostModel.
The parameter `C` defines the `cost_type` of the objective. The following
functions must be implemented for a `LowLevelCostModel` to be used:
* `get_initial_cost(model::LowLevelCostModel,env)` - returns
the initial_cost for a path
* `get_transition_cost(model::LowLevelCostModel{C},path::Path,s::S,a::A,
    sp::S) where {S,A,C}` - defines the cost associated with taking action
    `a` from state `s` to arrive in state `sp` according to the objective
    defined by `model` given that `s` is the "tip" of `path`.
* `accumulate_cost(model::LowLevelCostModel{C}, current_cost::C,
    transition_cost::C)` - defines how cost accumulates as new `PathNode`s
    are added to a Path.

CRCBS.LowLevelSolution — Type

`LowLevelSolution{S,A,T,C}`

Contains a list of agent paths and the associated costs.
Params:
- `S` is the state type
- `A` is the action type
- `T` is the cost type
- `C` is the cost model type
Elements:
- `paths::Vector{Path{S,A,T}}` is the vector of paths
- `costs::Vector{T}` is the vector of costs, one per path
- `cost::T` is the total cost for the entire solution

CRCBS.MAPF — Type

A MAPF is a Multi Agent Path Finding problem. It consists of an environment,
`env`, through which a group of agents may navigate, as well as a list of
start and goal states in that environment. Note that this is the _labeled_
case, where each agent has a specific assigned destination.

Elements:
- env::E - the base environment
- starts::Vector{S} - the vector of initial states
- starts::Vector{G} - the vector of goals

CRCBS.MetaAgentCBS_Solver — Type

MetaAgentCBS_Solver

Path planner that employs Meta Agent Conflict-Based Search

CRCBS.MetaCost — Type

`MetaCost`

`MetaCost` maintaining separate costs for individual agents that have been
combined into a MetaAgent.
- independent_costs::Vector{T} a vector of costs, 1 per agent
- total_cost::T the total cost, which reflects the combined cost (its
    interpretation depends on the `MetaCostModel` used to define cost-
    upating behavior)

CRCBS.MetaCostModel — Type

`MetaCostModel`

Defines the cost-updating behavior of `MetaCost` for MetaAgent applications.

CRCBS.MetaSolution — Type

MetaSolution{S}

Wrapper for a LowLevelSolution that allows for keeping track of groups.

CRCBS.MultiDeadlineCost — Type

`MultiDeadlineCost`

CRCBS.MultiStagePerfectHeuristic — Type

`MultiStagePerfectHeuristic`

Stores multiple lookup tables corresponding to different stages of a Path-
Finding search. Each stage has a different goal. The heuristic value at a
particular stage must reflect not just the distance to the next goal but the
length of the path through all remaining goals.

CRCBS.NullHeuristic — Type

`NullHeuristic`

CRCBS.PIBTCache — Type

PIBTCache{S,A}

Contains info to be passed along through recursive calls to the PIBT algorithm for multi-agent path planning. Info to be stored:

current state of each agent (should be lined up at the same time step)

priority of each agent

the planned action (and associated next state) for each agent

the search environment for each agent, which contains e.g., the agent's goal, costmodel, heuristicmodel, etc.

a conflict table of sorts to indicate which states/actions are reserved

countdown flags that identify which paths are "active". If pibt is operating

on a "ragged" plan, where some paths have been planned further into the future than others, it needs to ensure that planning does not continue for a given path until all of the other paths have "caught up" to it.

CRCBS.PIBTPlanner — Type

PIBTPlanner{C}

Planner based on Priority Inheritance with BackTracking.

Priority Inheritance with Backtracking for Iterative Multi-agent Path Finding Okumura et al, IJCAI 2019 https://www.ijcai.org/Proceedings/2019/0076.pdf

CRCBS.PIBTReservationTable — Type

PIBTReservationTable

CRCBS.Path — Type

`Path{S,A,C}`

Encodes a motion plan as a sequence of `PathNode{S,A}`s

CRCBS.PathNode — Type

`PathNode{S,A}`

Includes current state `s`, action `a`, next state `sp`

CRCBS.PerfectHeuristic — Type

`PerfectHeuristic`

The Perfect Heuristic stores the exact distance between any vertex in a
graph to all goal vertices specified during construction. The distances are
stored in `dists`, a dictionary which maps a goal vertex `v` to a vector of
distances from any other vertex in `1:nv(G)` to `v`.

An example of how to access the distance from vertex `v1` to goal `v2`:
`get_heuristic_cost(h,v1,v2) = h.dists[v2][v1]`

CRCBS.ReservationTable — Type

ReservationTable

Data structure for reserving resources over a time interval. The table stores a vector of reservations for each resource. When a new reservation is added to the table, it is inserted into the reservation vector associated to the requested resource.

CRCBS.ResourceReservation — Type

ResourceReservation

r::ResourceReservation encodes a that resource r.resource is reserved by agent r.agent_id over time interval r.interval.

CRCBS.SoftConflictTable — Type

`SoftConflictTable`

CRCBS.SoftConflictTable — Type

`construct_empty_lookup_table(graph,T::Int)`

Returns a soft lookup table to encode possible paths for each agent through graph. The argument T defines the time horizon of the lookup table.

CRCBS.SolverException — Type

SolverException

Custom exception type for tracking solver timeouts, etc.

CRCBS.SolverLogger — Type

SolverLogger

A logger type for keeping track of thing like runtime, iterations, optimality gap (including upper and lower bound), etc. The following methods allow for accessing and modifying the SolverLogger's fields, and are extended to solver types that wrap a SolverLogger:

iterations(logger) iterationlimit(logger) maxiterations(logger) starttime(logger) runtimelimit(logger) deadline(logger) JuMP.lowerbound(logger) bestcost(logger) verbosity(logger) verbosity(val::Int) debug(logger) solver_status(logger)

setiterations!(solver,val) incrementiterationcount!(logger::SolverLogger) setiterationlimit!(solver,val) setmaxiterations!(solver,val) setstarttime!(solver,val) setruntimelimit!(solver,val) setdeadline!(solver,val) setlowerbound!(solver,val) setlowerbound!(logger::SolverLogger{C},val::C) where {C} setlowerbound!(logger::SolverLogger{NTuple{N,T}},val::R) where {N,T<:Real,R<:Real} setlowerbound!(logger::SolverLogger{T},val::R) where {T<:Tuple,R<:Real} setbestcost!(solver,val) setbestcost!(logger::SolverLogger,val) setbestcost!(logger::SolverLogger{NTuple{N,T}},val::R) where {N,T<:Real,R<:Real} setbestcost!(logger::SolverLogger{T},val::R) where {T<:Tuple,R<:Real} setverbosity!(solver,val) setdebug!(solver,val)

The following methods facilitate control flow based on solver status. checktime(logger) enforcetimelimit!(logger) checkiterations(logger) enforceiterationlimit!(logger)

CRCBS.SolverWrapper — Type

SolverWrapper

An abstract type whose concrete instances must have a solver field.

CRCBS.a_star! — Method

a_star!(env,start_state)

A generic implementation of the [A* search algorithm](http://en.wikipedia.org/wiki/A%2A_search_algorithm)
that operates on an Environment and initial state.

args:
- `env::E <: AbstractLowLevelEnv`
- `start_state`

The following methods must be implemented:
- is_goal(env::E,s::S)
- check_termination_criteria(env::E,cost::C,path::Path{S,A,C},state::S)
- get_possible_actions(env::E,s::S)
- get_next_state(env::E,s::S,a::A,sp::S)
- get_transition_cost(env::E,s::S,a::A)
- violates_constraints(env::E,s::S,path::Path{S,A,C})

CRCBS.a_star_impl! — Method

The internal loop of the A* algorithm.

# g(n) = cost of the path from the start node to n,
# h(n) = heuristic estimate of cost from n to goal
# f(n) = g(n) + h(n)

CRCBS.accumulate_cost — Function

accumulate_cost(model,current_cost,transition_cost)

Defines the way that a transition_cost updates the current_cost of a Path.

CRCBS.action_space_trait — Method

action_space_trait(env)

Defaults to DiscreteSpace

CRCBS.add_conflict! — Method

helper to insert conflicts into ConflictTable

CRCBS.add_constraint! — Function

Adds a CBSConstraint to a DiscreteConstraintTable

CRCBS.add_constraint! — Method

adds a `CBSConstraint` to a ConstraintTreeNode

CRCBS.add_fat_path_to_table! — Function

`add_fat_path_to_table(CAT,fat_path)`

CRCBS.add_heuristic_cost — Function

`add_heuristic_cost(cost_model,heuristic_model,cost,h_cost)`

Defines the output heuristic cost that results from combining the current path cost cost with a heuristic "cost-to-go" h_cost. Gener

CRCBS.add_to_path! — Method

add_to_path!(path,env,s,a,sp)

Adds the new (s,a,sp) tuple and its cost (under env) to path.

CRCBS.aggregate_costs — Function

aggregate_costs(model, costs::Vector{T}) where {T}

Defines how costs from multiple distinct paths are combined into a single cost for the whole solution. For example, the aggregation function for a SumOfTravelTime objective is the sum of individual cost.

CRCBS.aggregate_costs_meta — Method

a special version of aggregate_costs for the meta_env

CRCBS.build_env — Function

`build_env(mapf::AbstractMAPF, node::ConstraintTreeNode, idx::Int)`

Constructs a new low-level search environment for a conflict-based search
mapf solver

CRCBS.cbs! — Method

Conflict-Based Search

Sharon et al 2012
https://www.aaai.org/ocs/index.php/AAAI/AAAI12/paper/viewFile/5062/5239

CRCBS.cbs_branch! — Method

cbs_branch!(solver,mapf,node,conflict,priority_queue)

Part of CBS interface. Defaults to splitting on the conflict and adding two nodes to the priority_queue, where each of the child nodes has one of the new complementary constraints.

CRCBS.cbs_bypass! — Method

cbs_bypass!(solver,mapf,node,conflict,priority_queue)

Part of CBS interface. Defaults to false, but can be overridden to modify the priority_queue and/or bypass the branching step of CBS

CRCBS.cbs_update_conflict_table! — Method

cbs_update_conflict_table!(solver,mapf,node,constraint)

Allows for flexible conflict-updating dispatch. This function is called within within the default cbs_branch!() method.

CRCBS.check_termination_criteria — Function

`check_termination_criteria(env::E <: AbstractLowLevelEnv{S,A,C}, cost,
    path::Path{S,A,C}, s::S)`

returns true if any termination criterion is satisfied

CRCBS.combine_agents! — Method

`combine_agents(conflict_table, groups::Vector{Vector{Int}})`

Helper for merging two (meta) agents into a meta-agent

CRCBS.compute_path_cost — Method

compute_path_cost(model,env,path,i)

Compute the cost of a path from scratch.

CRCBS.construct_config_dataframe — Method

construct_results_dataframe(loader,solver_config,config_template)

Compile results at solver_config.results_path into a DataFrame based on the features stored in solver_config.feats

CRCBS.construct_empty_lookup_table — Method

`construct_empty_lookup_table(G,T::Int)`

Returns an empty lookup table.

CRCBS.construct_results_dataframe — Method

construct_results_dataframe(loader,solver_config,config_template)

Compile results at solver_config.results_path into a DataFrame based on the features stored in solver_config.feats

CRCBS.convert_to_vertex_lists — Method

Helper for displaying Paths

CRCBS.count_conflicts — Method

`count_conflicts(conflict_table::ConflictTable,i::Int,j::Int)`

helper for counting the number of conflicts between agent i and agent j

CRCBS.create_reservations — Function

create_reservation(env,s,a,sp)

Must be overriden for environment env and the relevant state / action types.

CRCBS.create_reservations — Function

create_reservations(env::GraphEnv,s,a,sp,t=-1)

Generates three reservations as follows

t0 = get_t(s)
tF = get_t(sp)
t_mid = (t0+tF)/2
reservations = [
    ResourceReservation{Int}(s_idx,get_agent_id(env),   (t0,    t_mid)),
    ResourceReservation{Int}(a_idx,get_agent_id(env),   (t0,    tF)),
    ResourceReservation{Int}(sp_idx,get_agent_id(env),  (t_mid, tF)),
]

In this way, the reservations for one path node will not interfere with those for the next path node.

CRCBS.default_solution — Method

`default_solution(solver, mapf::AbstractMAPF)`

Defines what is returned by the solver in case of failure to find a feasible
solution.

CRCBS.deserialize — Function

deserialize(env,idx,t=-1)

Decodes an integer encoding of a state or action of type state_type(env) or action_type(env)

CRCBS.detect_action_conflict — Method

Checks for an ActionConflict between two Paths at time t

CRCBS.detect_action_conflict — Method

Detect an `ActionConflict` between two path nodes. Must be overridden for
each specific path class

CRCBS.detect_conflicts! — Method

add detected conflicts to conflict table

CRCBS.detect_conflicts! — Method

add detected conflicts to conflict table

CRCBS.detect_conflicts! — Method

detect conflicts between paths

CRCBS.detect_conflicts! — Method

Populates a `ConflictTable` with all conflicts that occur in a given vector
of paths. Conflict checking is performed in a pairwise fashion between
all paths.

args:
- conflict_table        a `ConflictTable` to store the detected conflicts
- paths:                a list of `Path`s, one for each individual agent
- idxs                  (optional) a list of agent ids for which to check
                        collisions against all other agents

CRCBS.detect_state_conflict — Method

Checks for a StateConflict between two Paths at time t

CRCBS.detect_state_conflict — Method

Detect a `StateConflict` between two path nodes. Must be overridden for each
specific path class

CRCBS.extend_path! — Method

extend_path!(path,T)

Extends `path` to match a given length `T` by adding `PathNode`s
corresponding to waiting at the final state.

args:
- path      the path to be extended
- the desired length of the new path

CRCBS.extend_path — Method

extend_path(path,T)

Extends a copy of `path` to match a given length `T` by adding `PathNode`s
corresponding to waiting at the final state.

args:
- path      the path to be extended
- the desired length of the new path

CRCBS.extract_feature — Function

extract_feature(solution,extractor,solution,solve_time)

Function must be implemented to extract features from solutions Args:

solver (pc_tapf solver)

extractor <: FeatureExtractor

solution <: SearchEnv

timer_results::NamedTuple{t,bytes,gctime,memallocs}

CRCBS.generate_constraints_from_conflict — Method

generates a Vector of (State or Action) Constraints from a conflict

CRCBS.get_agent_idxs — Method

get_agent_idxs(solver,node,mapf,constraint)

Part of CBS interface. Defaults to return the index of a single agent affected by a constraint. Can be overridden to return the index of e.g., a "meta-agent" (group of agents).

CRCBS.get_conflict_index — Method

get_conflict_index(cache,i,s,a,sp)

Returns the index of an agent that currently occupies sp, or -1 if there is no such agent.

CRCBS.get_conflicting_paths — Method

`get_conflicting_paths`

operates over a lookup table and returns a dictionary mapping path index to
the time index at which the conflict occurred

CRCBS.get_conflicts — Method

helper for retrieving conflicts associated with agents i and j

CRCBS.get_constraints — Method

retrieve constraints corresponding to this node and this path

CRCBS.get_cost_model — Method

Override this method for when the cost model has arguments

CRCBS.get_fat_path — Method

`get_fat_path(G,D,start_vtx,goal_vtx)`

returns a fat path through `G` from `start_vtx` to `goal_vtx`. Each set
of vertices in the fat path contains all vertices with distance d1 from
start_vtx and distance d2 to goal_vtx, where d1+d2 == the length of the
shortest path(s) from `start_vtx` to `goal_vtx`

G is a graph, D is the distance matrix

CRCBS.get_fat_path_cost — Method

get_fat_path_cost(model,nodes)

Returns a scalar cost value depending on typeof(model) and length(nodes).

get_fat_path_cost(m::FlatFPCost,nodes) = 1.0
get_fat_path_cost(m::NormalizedFPCost,nodes) = 1.0/length(nodes)

CRCBS.get_heuristic_cost — Function

`get_heuristic_cost(env::E <: AbstractLowLevelEnv{S,A,C},state::S)`

get a heuristic "cost-to-go" from `state`

CRCBS.get_infeasible_cost — Method

get_infeasible_cost(model)

Part of cost model interface. Defaults to zero.

CRCBS.get_infeasible_solution — Method

`get_infeasible_solution`

CRCBS.get_initial_cost — Method

get_initial_cost(model)

Part of cost model interface. Defaults to zero.

CRCBS.get_initial_solution — Method

`get_initial_solution`

CRCBS.get_level_set_nodes — Function

get_level_set_nodes(env,s,threshold,cost=get_initial_cost(env))

Returns a vector of PathNodes, where the heuristic cost (according to env) of each node falls below threshold.

CRCBS.get_next_conflict — Method

`get_next_conflict(conflict_table::ConflictTable)`

Returns the next conflict (temporally) that occurs in a conflict table

CRCBS.get_next_state — Function

`get_next_state(env::E <: AbstractLowLevelEnv{S,A,C}, s::S, a::A)`

returns a next state s

CRCBS.get_path_cost — Function

`get_path_cost(env::E <: AbstractLowLevelEnv{S,A,C},path::Path{S,A,C})`

get the cost associated with a search path so far

CRCBS.get_path_node — Method

returns the `PathNode` (s,a,s') corresponding to step `t` of `path`

If `t` is greater than the length of `path`, the `PathNode` returned
is (s,wait(s),s) corresponding to waiting at that node of the path.

path[t] is the path node that begins at t-1 and terminates at t

CRCBS.get_possible_actions — Function

`get_possible_actions(env::E <: AbstractLowLevelEnv{S,A,C}, s::S)`

return type must support iteration

CRCBS.get_transition_cost — Function

`get_transition_cost(env::E <: AbstractLowLevelEnv{S,A,C},s::S,a::A,sp::S)`

return scalar cost for transitioning from `s` to `sp` via `a`

CRCBS.handle_solver_exception — Method

handle_solver_exception

Takes care of printing SolverExceptions

CRCBS.hard_reset_solver! — Method

hard_reset_solver!(solver)

To be called when no information (other than iteration and time limits) needs to be stored.

CRCBS.has_constraint — Method

has_constraint(env,table,c::CBSConstraint)

CRCBS.init_dataframe — Function

init_dataframe(feats::Tuple)

Instantiate an empty dataframe based on the names and types of feature extractors in feats.

CRCBS.init_mapf_2 — Method

switch places

CRCBS.init_mapf_3 — Method

congested

CRCBS.init_mapf_4 — Method

almost switch corners. With the fat path heuristic, the paths should be:
- Robot 1: [1,5,9,13,14,15]
- Robot 2: [16,12,8,4,3,2]

CRCBS.init_mapf_5 — Method

PIBT demo from paper

CRCBS.initialize_child_search_node — Method

`initialize_child_search_node(parent_node::ConstraintTreeNode)`

Initialize a new `ConstraintTreeNode` with the same `solution` and
`constraints` as the parent node

CRCBS.initialize_root_node — Function

`initialize_root_node`

Construct an empty `ConstraintTreeNode` from a `AbstractMAPF` instance

CRCBS.is_available — Method

is_available

Returns false if the proposed reservation is not available to any of the agent_ids

CRCBS.is_consistent — Method

is_consistent(solution,mapf)

Check if solution satisfies start and end constraints

CRCBS.is_goal — Function

`is_goal(env,s)`

Returns true if state `s` satisfies the goal condition of environment `env`

CRCBS.is_valid — Method

Checks if a conflict is valid

CRCBS.load_problem — Function

load_problem(loader,filename)

Generic function that allows a problem instance to be retrived/constructed from loader (which may cache information) and filename::String (which points to a problem file).

CRCBS.low_level_search! — Method

`low_level_search!(
    solver,
    mapf::AbstractMAPF,
    node::ConstraintTreeNode,
    idxs=collect(1:num_agents(mapf)),
    path_finder=a_star!)`

Returns a low level solution for a MAPF with constraints. The heuristic
function for cost-to-go is user-defined and environment-specific

CRCBS.num_actions — Function

num_actions(env)

Returns the cardinality of the single agent state space for an environment. If the state and action spaces are finite and discrete, a discrete constraint table may be used for fast lookup.

CRCBS.num_states — Function

num_states(env)

Returns the cardinality of the single agent state space for an environment. If the state and action spaces are finite and discrete, a discrete constraint table may be used for fast lookup.

CRCBS.partially_set_path! — Method

`partially_set_path!`

Only replaces the cached path from start_time to length(path). Useful if you
want the remainder of the cached path to stay in the lookup table (i.e. for
repairing an existing plan).

CRCBS.pibt_priority_law — Method

pibt_priority_law(solver,mapf,cache,i)

Returns a value that will determine the priority of agent i relative to other agents. A lower value means higher priority.

CRCBS.pibt_step! — Function

pibt_step!(solver,mapf,i,j=-1)

i is the id of the higher priority agent, j is the index of the lower priority agent.

CRCBS.populate_soft_lookup_table! — Function

`populate_soft_lookup_table!(CAT,start_times,start_vtxs,goal_vtxs)`

CRCBS.recompute_cost — Function

recompute_cost

Recompute the cost of path (according to env), beginning from initial cost c0.

CRCBS.reconstruct_path! — Method

reconstruct path by working backward from the terminal state

CRCBS.remove_constraint! — Function

Removes a CBSConstraint to a DiscreteConstraintTable

CRCBS.reserve! — Method

reserve!(table,reservation)

Attempts to add a reservation to the table. If the reservation conflicts with an existing reservation, does nothing and returns false. If successful, returns true.

CRCBS.reserved_by — Method

reserved_by

Returns the IDs of agents who have reserved a resource within a specific time window.

CRCBS.reset_solver! — Method

reset_solver!(solver)

Resets iteration counts and start times.

CRCBS.reset_undecided! — Method

Fills `undecided` with all active agents (inactive agents have already
selected their actions)

CRCBS.run_profiling — Method

run_profiling(config,loader)

Profile the performance of one or more solvers on a set of problems defined by files in a directory. Args:

config: A named tuple or struct with at least the following fields:

problem_dir::String : path to problem files.

solver_configs : A vector of objects, each of which has the fields:

solver : a solver complying with the CRCBS interface.

results_path::String : path at which to store results.

feats : A vector of Feature objects, which defines the specific

features to be extracted and saved.

Loader: Any kind of cache on which load_problem(loader,problem_file) can be called.

CRCBS.search_constraints — Method

search_constraints(env,table,n::PathNode)

Returns all `CBSConstraint`s and `CBSConstraint`s that match `n`,
regardless of time.

CRCBS.serialize — Function

serialize(env,state,t=-1)

Encodes a state or action as an integer

CRCBS.set_constraint! — Function

Sets a `CBSConstraint` value in a DiscreteConstraintTable

CRCBS.solve! — Method

`solve!(solver, args ...)`

Run the algorithm represented by `solver` on an instance of a Multi-Agent
Path-Finding problem.

CRCBS.sorted_actions — Method

sorted_actions(env,s)

Return a vector of actions sorted lowest cost to highest cost.

CRCBS.state_space_trait — Method

state_space_trait(env)

Defaults to DiscreteSpace

CRCBS.states_match — Function

`state_match(s1::S,s2::S)`

returns true if s1 and s2 match (not necessarily the same as equal)

CRCBS.time_to_deadline — Method

time_to_deadline(solver)

time to deadline(solver) or runtime_limit(solver)–whichever is shorter.

CRCBS.trim_path! — Method

trim_path!(env,path,T)

Modify path to terminate at time step T. If length(path) < T, the path will be extended to time step T.

CRCBS.trim_solution! — Method

trim_solution!

Modify solution so that all paths terminate at time step T.

CRCBS.update_fat_path_conflict_table! — Method

update_conflict_table!(table,nodes)

Updates a conflict table with a set of nodes.

CRCBS.violates_constraints — Function

`violates_constraints(env::E <: AbstractLowLevelEnv{S,A,C},
    path::Path{S,A,C},s::S,a::A,sp::S)`

returns `true` if taking action `a` from state `s` violates any constraints
associated with `env`

CRCBS.wait — Function

`wait(s)`

returns an action that corresponds to waiting at state s

CRCBS.write_results — Method

write_results(loader,solver_config,prob,problem_file,results)

Write results results of profiling solver_config.solver on problem prob of name problem_name. Can be overloaded to dispatch on problem type when applicable.

GraphUtils.get_graph — Method

GraphUtils.get_graph(env::GraphEnv)

Must be implemented for all concrete subtypes of GraphEnv