CRCBS

CRCBS

CRCBS.GraphPath — Type.

Type alias for a path through the graph

source

CRCBS.traversal_time — Function.

returns the time for traversal of a GraphPath. Defaults to the length of the
path, but it may be useful in case we need to override later

source

CRCBS.LowLevelSolution — Type.

Type alias for a list of agent paths

source

CRCBS.MAPF — Type.

A MAPF is an instance of a Multi Agent Path Finding problem. It consists of
a graph `G` whose edges have unit length, as well as a set of start and goal
vertices on that graph. Note that this is the _labeled_ case, where each
agent has a specific assigned destination.

source

CRCBS.NodeConflict — Type.

Encodes a conflict wherein two agents occupy a particular node at a
particular time

source

CRCBS.invalid_node_conflict — Function.

Returns an invalid NodeConflict

source

CRCBS.EdgeConflict — Type.

Encodes a conflict between two agents at a particular edge at a particular
time. This means that the agents are trying to swap places at time t.

source

CRCBS.invalid_edge_conflict — Function.

Returns an invalid EdgeConflict

source

CRCBS.is_valid — Function.

checks if a solution is valid

source

Checks if a node conflict is valid

source

checks if an edge node is invalid

source

CRCBS.CBSConstraint — Type.

Encodes a constraint that agent `a` may not occupy vertex `v` at time `t`

source

CRCBS.ConstraintTreeNode — Type.

A node of a constraint tree. Each node has a set of constraints, a candidate
solution (set of robot paths), and a cost

source

CRCBS.ConstraintDict — Type.

constraint dictionary for fast constraint lookup within a_star

source